Remove chmod +x for installation script for docker build. (#13772)
We've introduced chmod a+x for installation scripts in Dockerfiles.
but this turned out to be a bad idea. This was to accomodate
building on Azure Deveops which has filesystem that does not
keep executable bit. But the side-effect of it that the
layer of the script is invalidated when the permission is changed
to +x on linux. The problem is that the script has locally (on
checkout) different permissions depending on umask setting.

Therefore changing permissions for the image to +a is not best.

Instead we are running the scripts with bash directly, which does
not require changing of executable bit.

(cherry picked from commit 1053310827d62871a8ee99662de2dff070ba790e)